Make the optional arguments to CallableType explicit
The CallableType constructor takes up to 11 arguments, which
is too many to read clearly as positional arguments.  Long
argument lists should be provided with names, both for
clarity and to avoid bugs if parameters are removed or new
ones inserted.

At the same time, many invocations pass only the 5 required
arguments, which are relatively manageable, and it might be
annoying to force explicitness on each of those.

Use the explicit names for the optional arguments wherever
we pass them.