[1/2] Fix creation of TypeVarType.
We weren't actually passing in the variance, instead
we were passing in line as the value and letting line
get defaulted.

As a possible fix for things like this not happening
in the future, invariance really should be an enum,
not an int. Otherwise, we don't get protected by the
type system we are making...