Refactor handling of arguments in parser.
I took a look at the function parse_arg_list and was appalled
by the fact that it was returning a 9-tuple of equal-length
lists.

This carries over into the Nodes; this patch does _not_ attempt
to go that far. It only consolidates into the 'Argument' class
during the parser, and then when CallableType or FuncDef are
created, they get split back into their parallel lists.

Getting the data initially parsed into the form we eventually
want to use it in should be the first step in making Argument
a real node used by FuncDef. (I think CallableType would actually
gain access to more data in the transition, since Argument contains
more than FuncDef does).

Eventually 'Argument' should be a node but I left it in parse.py
for now is it is not yet.