Allow multisource builds.
This is mostly a refactor but has a small change where we always
add os.cwdir() as a lib_path to make life more sane in a multisource
world.

Basically abstract all of the source types into a BuildSource
class and then take a list of them. Create multiple initial states,
being careful that we don't load their dependencies until after
all of them are added to the states (if we do this in the wrong order,
the dependencies may add one of the remaining states-to-be-added).