Only print test output on failures or -v.
First, we shouldn't have multiple processes outputting to the same
stdout/stderr filehandles. This made the output completely wonky and
difficult to interpret. Instead use subprocess.PIPE to make a new
stdout, and use subprocess.STDOUT to redirect the subprocesses' stderr
to stdout. Since we don't know if we are going to print these until the
process exits, we need to combine them so as to not lose their ordering.

If the process failed or we are running in verbose mode, only then do we
print the output. Because all prints to real stdout are from the main
thread, this removes the ability for all of the different processes to
interleave their output in weird ways.