Don't export imported names from stubs unless using 'as'
Now 'from m import *' does not implicitly import all names imported in
'm' if 'm' has a stub file.

This only deals with 'from ... import' in a stub file.

I changed the type of names in ImportFrom to have the 'as id'
attribute None if there is no explicit 'as', which resulted in
most of the changes.