Cache most searches for modules through the path
In type-checking one codebase of about 50 files and 150 direct
dependencies with --silent, we were spending 4.0s out of 8.2s
in find_module, of which 3.8s was in is_module, and the bulk of
those were duplicate calls.

Adding this cache cut out 89% of the total time spent in
find_module, and 44% of all the time spent in the whole program.
Now find_module is just under 10% of the time.

We could probably gain another few percent by caching other
calls to find_module too.  This is more complicated because
we don't necessarily always use the same lib_path -- leave
it for a subsequent diff.