Fix tests for find_module caching
In unit tests, we call `build.build` many times over the lifetime of
the same Python process, with the same `lib_path` but having mucked
around in the filesystem in between so that our cached picture of the
files in lib_path is no longer accurate.  We should keep these caches
only within a type-checking run instead.

Also, when passed a `lib_path` entry with a trailing slash, the new
code was ending up with a double-slash in the result returned from
`find_module`.  Harmless but a little unsightly, and more annoyingly
it doesn't match the expected output in some tests.  So clean that up.