runtests: rationalize add_mypy_* a bit more
Put them in one place.  Cut the duplicate name logic.  Cut the
add_both_* methods that nothing uses.  Fix the add_mypy_mod interface
that would never work if actually fully used, and make a version that
would.  Cut the add_mypy_mod callsites that are redundant with the
self-typecheck, and update the one remaining callsite to the new interface.