Generate correct constraints for Unions (#1408)
* Simplify unions when expanding types

* Generate correct constraints for Unions

The old code always generated a conjunction, where we sometimes want a
disjunction.  Since we can't represent or solve a disjunction of
constraints, need to approximate.

In the presence of Unions, the constraints inferred from (template,
actual, direction) are not necessarily the reverses of the constraints
inferred from (template, actual, neg_op(direction)). The
visit_instance and visit_callable cases were updated accordingly. Test
testGenericFunctionSubtypingWithUnions fails without this change.

Fixes #1458 and the remaining part of #1241.