Pr/strict optional (#1562)
First pass strict Optional checking.  Adds the experimental `--strict-optional`
flag. Fixes #1450.

With --strict-optional:
- "None" in type annotations refers to NoneTyp, except as a return
  value, where it still refers to Void
- None and List[None] will now be inferred as a type if no other
  information is available.
- Class variables may be initialized to None without having an Optional
  type.  Mypy does not currently check that they're assigned to in
  __init__ or elsewhere before use.
See #1450 for more details.

This also fixes the bug where mypy didn't understand that x.y = "foo"
implied that x.y would be a str for the remaineder of that block.  This
isn't entirely sound, but is in line with the way we do isinstance
checks.