Improve handling of "lambda: None" (#1557)
"lambda: None" might be used either where a function returning an
Optional[T] is expected, or where a function not returning a value is
expected. Previously it would always treated as the former; now the
context is used to decide.

When no context is available, mypy continues to treat the lambda as
returning the value None, rather than no value at all. When the other
meaning is desired, it is generally possible to provide a context, or
in any case to use a function defined with `def`.

Fixes #1425.