Fix crash when constructor annotation is too short.
This was reported internally and I have not been able to find a short
repro.  The pattern was something like this:
```
class C:
  def __init__(self, a=None, b=None):
    # type: (int) -> None
    print(a, b)

def f():
  # type: () -> C
  return C(a=1)
```
However, that doesn't crash, so there's more to it (possibly an import
cycle), and I'm out of time researching this.

I've verified that this fix avoids the crash, and it feels pretty
reasonable to me. Unless it breaks some test I intend to merge it
right away.