Set up to use pytest for our data-driven tests, and switch testcheck over to it (#1944)
This is a step toward #1673 (switching entirely to pytest from myunit
and runtests.py), using some of the ideas developed in @kirbyfan64's
PR #1723.

Both `py.test` with no arguments and `py.test mypy/test/testcheck.py`
work just as you'd hope, while `./runtests.py` continues to run all
the tests.

The output is very similar to the myunit output.  It doesn't spew
voluminous stack traces or other verbose data, and it continues using
`assert_string_arrays_equal` to produce nicely-formatted comparisons
when e.g. type-checker error messages differ.  On error it even
includes the filename and line number for the test case itself, which
I've long wanted, and I think pytest's separator lines and coloration
make the output slightly easier to read when there are multiple
failures.

The `-i` option from myunit is straightforwardly ported over as
`--update-data`, giving it a longer name because it feels like the
kind of heavyweight and uncommon operation that deserves such a name.
It'd be equally straightforward to port over `-u`, but in the presence
of source control I think `--update-data` does the job on its own.

One small annoyance is that if there's a failure early in a test run,
pytest doesn't print the detailed report on that failure until the
whole run is over.  This has often annoyed me in using pytest on other
projects; useful workarounds include passing `-x` to make it stop at
the first failure, `-k` to filter the set of tests to be run, or
(especially with our tests where errors often go through
`assert_string_arrays_equal`) `-s` to let stdout and stderr pass
through immediately.  For interactive use I think it'd nearly always
be preferable to do what myunit does by immediately printing the
detailed information, so I may come back to this later to try to get
pytest to do that.

We don't yet take advantage of `xdist` to parallelize within a
`py.test` run (though `xdist` works for me out of the box in initial
cursory testing.)  For now we just stick with the `runtests.py`
parallelization, so we set up a separate `py.test` command for each
test module.