Add new --platform flag (#2045)
* Add cmd line flag to let you set the OS version

This adds a new flag allowing you to set what the expected value of
sys.platform is for this particular run. It is similar to the
--python-version flag, which lets you control what the expected value
of sys.version_info for this run is.

* Support sys.platform.startswith(...) checks

Previously, mypy only took into account sys.platform of the form
`sys.platform == ???` or `sys.platform != ???`.

This adds support for `sys.platform.startswith(???)` checks. This will
probably be helpful when writing Python2/Python3-straddling code. For
example, Python 2, sys.platform returns 'linux2' on Linux systems, but
in Python 3, sys.platform will return just 'linux'.

Fixes #1988.