Modify cache file formats to shrink overall cache size (#2108)
* Compress usually-false flags when serializing

This commit modifies several mypy nodes that tend to contain a large
number of usually-false boolean flags to store a compressed version of
those flags.

More precisely, rather then storing each flag and its boolean value as a
key-value pair in the output JSON, we now store a list of the flags that
are known to be true at time of serialization.

When deserializing, we iterate over the list of flags and set each
attribute to True, overriding the default value of False.

This does not result in any significant speedups, but does end up
reducing the size of the cache by a non-trivial factor.

* Stop serializing Argument.type_annotation

This commit modifies the serialization logic of Argument to stop saving
the type_annotation data.

After some investigation, it appears as if that data is never actually used
once the cache data is loaded by mypy, and removing it would help shrink
the cache size by a respectable margin.

* Stop caching all arguments in serialized FuncDefs

This commit modifies the serialization logic and some of the surrounding
code for FuncDefs so we no longer serialize the full argument list --
instead, we just stick to serializing the argument names and the
argument kinds.

As with the other commits, this does not really improve speed, but does
end up shrinking the cache size by a fair amount.

It would have been nice to avoid storing the arg kinds and arg names
altogether since the information can usually be found within the
corresponding callable type, but unfortunately, a FuncDef is not
guaranteed to actually have a type in all instances -- sometimes, the
type is None. This is a problem since we need the argument names and
kinds to be able to infer a fallback type later on in mypy.