Fix meta level on member access (#2809)
Fixes #2804

The problem is that analyze_member_access calls freshen_function_type_vars 
if the member is a function (i.e. method). This is necessary to not mix the type 
variables during type inference of generic methods inside generic functions. 
However, if the method does not participate in type inference, type variables 
are left in meta_level = 1 state. This causes the error, since the types could not 
be serialized in the middle of type inference.

In this PR I propose to restore meta_level = 0 on member access, this is safe, 
since anyway, check_call always calls freshen_function_type_vars on callee (for 
generic functions, generic methods, etc).