Translate f-string expression to str.format expression. (#2962)
This changes the way f-string expressions are translated. f-string
expressions are now translated into a str.format call. This allows
using expressions inside an f-string that do not have a callable
__str__ method (such as types). This commit also adds a test-case
where a type is used in an f-string expression and adds a .format
on the str primitive fixture. This fixes python/mypy#2947.