Remove bundled lib-typing (#3337)
This gets rid of code duplication and makes sure mypy works with the released version of `typing`. This doesn't fix any immediate issue, but I saw some comments in the bug tracker saying that we should be able to get rid of `lib-typing` eventually.

There is a minor downside: To run mypy tests locally, you will now have to have a Python 2.7 installed that has `typing` available.