Improve test output when a test doesn't specify the correct fixture (#3488)
Improve things in two ways:

1) Don't crash in tests due to missing types.ModuleType

The fixtures used in many test cases don't import the `types` module
to speed up tests, and this would often cause uncaught exceptions while
writing tests.  Now we generate a useful message instead.

2) Suggest fixtures

If a developer is doing certain common things in a mypy test that
require a non-default builtins fixture (such as `typing.List`), add a
hint that points to a fixture that is likely to help.