Add basic support for user-defined mypy plugins (#3517)
Configure them through "plugins=path/plugin.py, ..." in the ini file.
The paths are relative to the configuration file.

This is an almost minimal implementation and some features
are missing:

* Plugins installed through pip aren't properly supported.
* Plugins within packages aren't properly supported.
* Incremental mode doesn't invalidate cache files when
  plugins change.

Also change path normalization in test cases in Windows.
Previously we sometimes normalized to Windows paths and sometimes
to Linux paths. Now switching to always use Linux paths.