Make binder handle unconditional ifs with no else (#3567)
This commit fixes #2673.

The problem was that in the semantic analysis phase, we can only ever
mark blocks as *unreachable*. So, if we have code of the form:

```python
if typing.TYPE_CHECKING:
    assert isinstance(...)
```

...nothing happens since there's no "else" block to mark as unreachable,
which means the binder sees nothing special about this if statement and
ignores the special conditional.

I decided to just add a dummy empty "else" block if one doesn't already
exist and mark it as unreachable. Two other solutions I considered were
to add another field somewhere to record reachability or to repeat the
special conditional checks in the typechecking phase, but these
solutions seemed a bit redundant/wasteful.