Redesign plugins and add additional plugin hooks (#3534)
Instead of passing several arguments to hook function, always
pass just a single object. This simplifies the signatures of
hooks.

Instead of passing callback functions to hooks, pass an object
that implements a specific interface.

These changes are intended to make it easier to write plugins,
and to make it easier to evolve the plugin system. Adding
extra attributes to context or extra methods to the internal
interfaces doesn't require changes to existing plugins.

Additional updates:

* Add hook that overrides the inferred type of an instance attribute.
  In particular, this can be used to override the type of `__call__`.
* Add hook for custom semantic analysis of types. This makes it possible
  to have some limited custom syntax for generic types. (We can extend
  the syntactic possibilities in the future.)
* Use the name of a callable to decide which hook to apply. This makes
  it possible to use custom hooks for callables returned by functions.